WIP: Remove `ResumeTy` from async lowering
Instead of using the stdlib supported `ResumeTy`, which is being converting to a `&mut Context<'_>` during the Generator MIR pass,
this will use `&mut Context<'_>` directly in HIR lowering.

It pretty much reverts #105977 and re-applies an updated version of #105250.

This still fails the testcase added in #106264 however, for reasons I don’t understand.