Print thread ID in panic message if thread name is unknown
`panic!` does not print any identifying information for threads that are
unnamed. However, in many cases, the thread ID can be determined.

This changes the panic message from something like this:

    thread '<unnamed>' panicked at src/main.rs:3:5:
    explicit panic

To something like this:

    thread '<unnamed>' (1234) panicked at src/main.rs:3:5:
    explicit panic

This change applies to both named and unnamed threads. There is no
change for threads where an ID cannot be determined.