Assert that the first `assert!` expression is `bool`
In the desugaring of `assert!`, assign the condition expression to a `bool`
biding in order to provide better type errors when passed the wrong thing.

The span will point only at the expression, and not the whole `assert!`
invokation.

```
error[E0308]: mismatched types
  --> $DIR/issue-14091.rs:2:13
   |
LL |     assert!(1,1);
   |             ^
   |             |
   |             expected `bool`, found integer
   |             expected due to this
```

We no longer mention the expression needing to implement the `Not` trait.

```
error[E0308]: mismatched types
  --> $DIR/issue-14091-2.rs:15:13
   |
LL |     assert!(x, x);
   |             ^
   |             |
   |             expected `bool`, found `BytePos`
   |             expected due to this
```

Fix #122159.