support multiple stability attributes (a mix of stable and unstable)
An item is unstable if it has any unstable attributes, to make it easier to track which features
library items depended on as they stabilize.

This changes the text for E0544. Unfortunately, the example doesn't make much sense anymore.

The way this merges stability levels together is made to work for stability-checking and rustdoc;
as far as I can tell, only `rustc_passes::lib_features` needs them separate, and it extracts them
itself.